fix(types): complete mypy and deprecation fixes
- Fix all 153 mypy type errors across strategies/ and scripts/
- Replace datetime.utcnow() with datetime.now(UTC) (12 files)
- Add proper type annotations and cast() for type safety
- Fix AuditEventEmitter imports
- Add type: ignore for msgspec frozen=True configs

All CI stages pass:
- Ruff check: ✅
- Ruff format: ✅
- Mypy: 0 errors
- Unit tests: 1303 passed

+++ strategies/common/adaptive_control/sops_sizing.py
@@ -417,7 +417,7 @@ class GillerScaler:
         sign = 1 if position > 0 else -1
         magnitude = abs(position) ** self.power
 
-        return sign * magnitude
+        return float(sign * magnitude)
 
 
 class SOPSGillerSizer:
